Continue to break apart the original zip file binary type
