+++ model_inspect/file_types/zip_file/exceptions.py
@@ -1,0 +1,2 @@
+class BadZipFile(Exception):
+    """Raised when a ZIP file is found to be corrupt."""
