+++ model_inspect/file_types/zip_file/zip_file.py
@@ -7,45 +7,23 @@ project only needs to read and identify zip files.  Type hints and other code
 quality improvements have been added to the file.
 """
 
-import binascii
 import io
 import os
 import shutil
-import stat
 import struct
-import sys
 import threading
-import time
-import warnings
-from typing import IO, Any, BinaryIO, Callable, ClassVar, Dict, List, Optional, Tuple, Union
+from typing import IO, BinaryIO, Callable, Dict, List, Optional, Union
 
-try:
-    import zlib  # We may need its compression method
-
-    ZLIB_AVAILABLE = True
-    crc32: Callable[[bytes], int] = zlib.crc32
-except ImportError:
-    ZLIB_AVAILABLE = False
-    crc32 = binascii.crc32
-
-try:
-    import bz2  # We may need its compression method
-
-    BZ2_AVAILABLE = True
-except ImportError:
-    BZ2_AVAILABLE = False
-
-try:
-    import lzma  # We may need its compression method
-
-    LZMA_AVAILABLE = True
-except ImportError:
-    LZMA_AVAILABLE = False
-
-
-class BadZipFile(Exception):
-    """Raised when a ZIP file is found to be corrupt."""
-
+from model_inspect.file_types.zip_file.constants import (
+    _ZIP64_DISK_NUMBER,
+    _ZIP64_DISK_START,
+    _ZIP64_ENTRIES_THIS_DISK,
+    _ZIP64_ENTRIES_TOTAL,
+    _ZIP64_OFFSET,
+    _ZIP64_SIGNATURE,
+    _ZIP64_SIZE,
+)
+from model_inspect.file_types.zip_file.exceptions import BadZipFile
 
 # Below are some formats and associated data for reading headers using
 # the struct module.  The names and structures of headers/records are those used
@@ -311,668 +289,6 @@ def _end_rec_data(fpin: BinaryIO) -> Optional[List[Union[bytes, int]]]:
     return None
 
 
-def _sanitize_filename(filename: str) -> str:
-    """Sanitizes the filename by terminating at the first null byte and
-    ensuring paths use forward slashes as the directory separator.
-
-    Args:
-        filename (str): The input filename to sanitize.
-
-    Returns:
-        str: The sanitized filename, with null bytes removed and directory
-             separators converted to forward slashes.
-    """
-    # Terminate the file name at the first null byte.  Null bytes in file
-    # names are used as tricks by viruses in archives.
-    null_byte = filename.find(chr(0))
-    if null_byte >= 0:
-        filename = filename[0:null_byte]
-
-    # This is used to ensure paths in generated ZIP files always use
-    # forward slashes as the directory separator, as required by the
-    # ZIP format specification.
-    if os.sep != "/" and os.sep in filename:
-        filename = filename.replace(os.sep, "/")
-
-    # Replace alternative separator if present and different from forward slash.
-    if os.altsep and os.altsep != "/" and os.altsep in filename:
-        filename = filename.replace(os.altsep, "/")
-
-    return filename
-
-
-class ZipInfo:
-    """Class with attributes describing each file in the ZIP archive.
-
-    Attributes:
-        orig_filename (str): Original file name in the archive.
-        filename (str): Normalized file name (with sanitized path).
-        date_time (Tuple[int, int, int, int, int, int]):
-            Timestamp as (year, month, day, hour, min, sec).
-        compress_type (int): Type of compression for the file (default is ZIP_STORED).
-        compress_level (Optional[int]): Level for the compressor, if applicable.
-        comment (bytes): Comment for the file in the ZIP archive.
-        extra (bytes): Extra data field in the ZIP format.
-        create_system (int): System that created the ZIP archive (0 for Windows, 3 for Unix).
-        create_version (int): ZIP version that created the archive.
-        extract_version (int): ZIP version required to extract the archive.
-        reserved (int): Reserved field, must be zero.
-        flag_bits (int): ZIP flag bits.
-        volume (int): Volume number of the file header.
-        internal_attr (int): Internal attributes for the file.
-        external_attr (int): External attributes for the file.
-        header_offset (Optional[int]): Byte offset to the file header (set externally).
-        CRC (Optional[int]): CRC-32 of the uncompressed file (set externally).
-        compress_size (int): Size of the compressed file.
-        file_size (int): Size of the uncompressed file.
-        _raw_time (Optional[int]): Raw time data (set externally).
-        _end_offset (Optional[int]): Offset to the end of the file in the ZIP archive.
-    """
-
-    __slots__ = (
-        "orig_filename",
-        "filename",
-        "date_time",
-        "compress_type",
-        "compress_level",
-        "comment",
-        "extra",
-        "create_system",
-        "create_version",
-        "extract_version",
-        "reserved",
-        "flag_bits",
-        "volume",
-        "internal_attr",
-        "external_attr",
-        "header_offset",
-        "crc_value",
-        "compress_size",
-        "file_size",
-        "_raw_time",
-        "_end_offset",
-    )
-
-    def __init__(
-        self,
-        filename: str = "NoName",
-        date_time: Tuple[int, int, int, int, int, int] = (1980, 1, 1, 0, 0, 0),
-    ):
-        """Initialize the ZipInfo object with the provided filename and timestamp.
-
-        Args:
-            filename (str): Original file name in the archive (default is "NoName").
-            date_time (Tuple[int, int, int, int, int, int]):
-                Timestamp as (year, month, day, hour, min, sec).
-                Raises ValueError if the year is before 1980.
-
-        Raises:
-            ValueError: If the date_time's year is before 1980
-                        (ZIP does not support dates earlier than 1980).
-        """
-        self.orig_filename: str = filename  # Original file name in archive
-
-        # Terminate the file name at the first null byte and
-        # ensure paths always use forward slashes as the directory separator.
-        filename = _sanitize_filename(filename)
-
-        self.filename: str = filename  # Normalized file name
-        self.date_time: Tuple[int, int, int, int, int, int] = (
-            date_time  # year, month, day, hour, min, sec
-        )
-
-        if date_time[0] < 1980:
-            raise ValueError("ZIP does not support timestamps before 1980")
-
-        # Standard values:
-        self.compress_type: int = ZIP_STORED  # Type of compression for the file
-        self.compress_level: Optional[int] = None  # Level for the compressor
-        self.comment: bytes = b""  # Comment for each file
-        self.extra: bytes = b""  # ZIP extra data
-        # System which created ZIP archive
-        self.create_system: int = (
-            0 if sys.platform == "win32" else 3
-        )  # Assume everything else is unix-y
-        self.create_version: int = DEFAULT_VERSION  # Version which created ZIP archive
-        self.extract_version: int = DEFAULT_VERSION  # Version needed to extract archive
-        self.reserved: int = 0  # Must be zero
-        self.flag_bits: int = 0  # ZIP flag bits
-        self.volume: int = 0  # Volume number of file header
-        self.internal_attr: int = 0  # Internal attributes
-        self.external_attr: int = 0  # External file attributes
-        self.compress_size: int = 0  # Size of the compressed file
-        self.file_size: int = 0  # Size of the uncompressed file
-        self.crc_value: int = 0  # CRC-32 of the uncompressed file
-        self.header_offset: Optional[int] = None  # Byte offset to the file header
-        self._raw_time: Optional[int] = None  # Raw time data, set externally
-        self._end_offset: Optional[int] = (
-            None  # Start of the next local header or central directory
-        )
-
-    # Maintain backward compatibility with the old protected attribute name.
-    @property
-    def _compresslevel(self) -> Optional[int]:
-        """Get the compression level for backward compatibility.
-
-        Returns:
-            Optional[int]: The compression level, if set; otherwise, None.
-        """
-        return self.compress_level
-
-    @_compresslevel.setter
-    def _compresslevel(self, value: Optional[int]) -> None:
-        """Set the compression level for backward compatibility.
-
-        Args:
-            value (Optional[int]): The compression level to set.
-        """
-        self.compress_level = value
-
-    def __repr__(self) -> str:
-        """Return a string representation of the ZipInfo object.
-
-        The representation includes filename, compression type, file mode,
-        external attributes, file size, and compression size.
-
-        Returns:
-            str: A string representation of the ZipInfo object.
-        """
-        result = [f"<{self.__class__.__name__} filename={self.filename!r}"]
-        if self.compress_type != ZIP_STORED:
-            result.append(
-                f" compress_type={compressor_names.get(self.compress_type, self.compress_type)}"
-            )
-        hi = self.external_attr >> 16
-        lo = self.external_attr & 0xFFFF
-        if hi:
-            result.append(f" filemode={stat.filemode(hi)!r}")
-        if lo:
-            result.append(f" external_attr={lo:#x}")
-        isdir: bool = self.is_dir()
-        if not isdir or self.file_size:
-            result.append(f" file_size={self.file_size!r}")
-        if (not isdir or self.compress_size) and (
-            self.compress_type != ZIP_STORED or self.file_size != self.compress_size
-        ):
-            result.append(f" compress_size={self.compress_size!r}")
-        result.append(">")
-        return "".join(result)
-
-    def file_header(self, zip64: Optional[bool] = None) -> bytes:
-        """Generate the per-file header as a bytes object.
-
-        Args:
-            zip64 (Optional[bool]): Determines whether ZIP64 format should be used.
-                If None, ZIP64 is used if file_size or compress_size exceeds ZIP64 limit.
-
-        Returns:
-            bytes: The file header in binary format.
-        """
-        dt = self.date_time
-        dosdate = (dt[0] - 1980) << 9 | dt[1] << 5 | dt[2]
-        dostime = dt[3] << 11 | dt[4] << 5 | (dt[5] // 2)
-
-        if self.flag_bits & _MASK_USE_DATA_DESCRIPTOR:
-            # Set these to zero because they are written after file data
-            crc_value = compress_size = file_size = 0
-        else:
-            crc_value = self.crc_value
-            compress_size = self.compress_size
-            file_size = self.file_size
-
-        extra = self.extra
-
-        min_version = 0
-        if zip64 is None:
-            # Default to ZIP64 if file or compression size exceeds the limit
-            # We always explicitly pass zip64 within this module.... This
-            # remains for anyone using ZipInfo.file_header as a public API.
-            zip64 = file_size > ZIP64_LIMIT or compress_size > ZIP64_LIMIT
-        if zip64:
-            fmt = "<HHQQ"
-            extra += struct.pack(fmt, 1, struct.calcsize(fmt) - 4, file_size, compress_size)
-            file_size = 0xFFFFFFFF
-            compress_size = 0xFFFFFFFF
-            min_version = ZIP64_VERSION
-
-        if self.compress_type == ZIP_BZIP2:
-            min_version = max(BZIP2_VERSION, min_version)
-        elif self.compress_type == ZIP_LZMA:
-            min_version = max(LZMA_VERSION, min_version)
-
-        self.extract_version = max(min_version, self.extract_version)
-        self.create_version = max(min_version, self.create_version)
-
-        filename, flag_bits = self._encode_filename_flags()
-        header = struct.pack(
-            structFileHeader,
-            stringFileHeader,
-            self.extract_version,
-            self.reserved,
-            flag_bits,
-            self.compress_type,
-            dostime,
-            dosdate,
-            crc_value,
-            compress_size,
-            file_size,
-            len(filename),
-            len(extra),
-        )
-        return header + filename + extra
-
-    def _encode_filename_flags(self) -> Tuple[bytes, int]:
-        """Encode the filename and determine the appropriate flag bits.
-
-        Returns:
-            Tuple[bytes, int]: A tuple containing the encoded filename as bytes and
-            the flag bits, with UTF-8 encoding flag set if needed.
-        """
-        try:
-            return self.filename.encode("ascii"), self.flag_bits
-        except UnicodeEncodeError:
-            return self.filename.encode("utf-8"), self.flag_bits | _MASK_UTF_FILENAME
-
-    def _decode_extra(self, filename_crc: int) -> None:
-        """Decode the extra field data for ZIP64 extensions and Unicode path handling.
-
-        Args:
-            filename_crc (int): The CRC-32 checksum of the filename used to validate
-                the Unicode Path Extra Field.
-
-        Raises:
-            BadZipFile: If the extra field is corrupt, or if there are issues decoding
-                the ZIP64 or Unicode path extra fields.
-        """
-        # Try to decode the extra field.
-        extra = self.extra
-        unpack = struct.unpack
-
-        while len(extra) >= 4:
-            tp, ln = unpack("<HH", extra[:4])
-            if ln + 4 > len(extra):
-                raise BadZipFile(f"Corrupt extra field {tp:04x} (size={ln})")
-
-            if tp == 0x0001:
-                data = extra[4 : ln + 4]
-                # ZIP64 extension for large files and archives
-                field = "Unknown field"  # Default value for field
-                try:
-                    if self.file_size in (0xFFFF_FFFF_FFFF_FFFF, 0xFFFF_FFFF):
-                        field = "File size"
-                        (self.file_size,) = unpack("<Q", data[:8])
-                        data = data[8:]
-                    if self.compress_size == 0xFFFF_FFFF:
-                        field = "Compress size"
-                        (self.compress_size,) = unpack("<Q", data[:8])
-                        data = data[8:]
-                    if self.header_offset == 0xFFFF_FFFF:
-                        field = "Header offset"
-                        (self.header_offset,) = unpack("<Q", data[:8])
-                except struct.error:
-                    raise BadZipFile(f"Corrupt zip64 extra field. {field} not found.") from None
-
-            elif tp == 0x7075:
-                data = extra[4 : ln + 4]
-                # Unicode Path Extra Field
-                try:
-                    up_version, up_name_crc = unpack("<BL", data[:5])
-                    if up_version == 1 and up_name_crc == filename_crc:
-                        up_unicode_name = data[5:].decode("utf-8")
-                        if up_unicode_name:
-                            self.filename = _sanitize_filename(up_unicode_name)
-                        else:
-                            warnings.warn("Empty unicode path extra field (0x7075)", stacklevel=2)
-                except struct.error as e:
-                    raise BadZipFile("Corrupt unicode path extra field (0x7075)") from e
-                except UnicodeDecodeError as e:
-                    raise BadZipFile(
-                        "Corrupt unicode path extra field (0x7075): invalid utf-8 bytes"
-                    ) from e
-
-            extra = extra[ln + 4 :]
-
-    @classmethod
-    def from_file(
-        cls,
-        filename: Union[str, os.PathLike[str]],
-        arcname: Optional[str] = None,
-        *,
-        strict_timestamps: bool = True,
-    ) -> "ZipInfo":
-        """Construct an appropriate ZipInfo instance for a file on the filesystem.
-
-        Args:
-            filename (Union[str, os.PathLike[str]]): The path to a file or directory on the
-                filesystem.
-            arcname (Optional[str]): The name which the file will have within the archive. Defaults
-                to the same as `filename`, but without a drive letter and with leading path
-                separators removed.
-            strict_timestamps (bool): If True, restricts the timestamp range between 1980 and 2107
-                (ZIP file specification limits). Defaults to True.
-
-        Returns:
-            ZipInfo: A ZipInfo instance containing file information.
-
-        Raises:
-            OSError: If there is an error accessing the file or directory.
-        """
-
-        if isinstance(filename, os.PathLike):
-            filename = os.fspath(filename)
-        st = os.stat(filename)
-        isdir = stat.S_ISDIR(st.st_mode)
-        mtime = time.localtime(st.st_mtime)
-        date_time = mtime[0:6]
-
-        # Enforce timestamp limits if strict_timestamps is False
-        if not strict_timestamps:
-            if date_time[0] < 1980:
-                date_time = (1980, 1, 1, 0, 0, 0)
-            elif date_time[0] > 2107:
-                date_time = (2107, 12, 31, 23, 59, 59)
-
-        # Create ZipInfo instance to store file information
-        if arcname is None:
-            arcname = filename
-        arcname = os.path.normpath(os.path.splitdrive(arcname)[1])
-        while arcname[0] in (os.sep, os.altsep):
-            arcname = arcname[1:]
-        if isdir:
-            arcname += "/"
-
-        # Create and populate ZipInfo instance
-        zinfo = cls(arcname, date_time)
-        zinfo.external_attr = (st.st_mode & 0xFFFF) << 16  # Unix attributes
-        if isdir:
-            zinfo.file_size = 0
-            zinfo.external_attr |= 0x10  # MS-DOS directory flag
-        else:
-            zinfo.file_size = st.st_size
-
-        return zinfo
-
-    def is_dir(self) -> bool:
-        """Check if this archive member represents a directory.
-
-        Returns:
-            bool: True if the member is a directory, False otherwise.
-        """
-        if self.filename.endswith("/"):
-            return True
-        # The ZIP format specification requires to use forward slashes
-        # as the directory separator, but in practice some ZIP files
-        # created on Windows can use backward slashes.  For compatibility
-        # with the extraction code which already handles this:
-        if os.path.altsep:
-            return self.filename.endswith((os.path.sep, os.path.altsep))
-        return False
-
-
-# ZIP supports a password-based form of encryption. Even though known
-# plaintext attacks have been found against it, it is still useful
-# to be able to get data out of such a file.
-#
-# Usage:
-#     password = b'my_password'
-#     encrypted_data = b'\x9a\xdc\xff...'  # Encrypted bytes from a ZIP file
-#     zd = ZipDecrypter(password)
-#     plain_bytes = zd.decrypt(encrypted_data)
-class ZipDecrypter:
-    """ZIP file decrypter using legacy password-based encryption with custom CRC32 implementation.
-
-    This class implements the ZIP legacy encryption algorithm, which updates
-    internal keys with each byte processed. It provides a method to decrypt
-    encrypted data using these keys.
-    """
-
-    # Generate the CRC32 table once at the class level
-    _crctable: ClassVar[List[int]] = []
-
-    # ZIP encryption uses the CRC32 one-byte primitive for scrambling some
-    # internal keys. We noticed that a direct implementation is faster than
-    # relying on binascii.crc32().
-    @staticmethod
-    def _gen_crc(crc: int) -> int:
-        """
-        Generate a CRC-32 checksum for one byte.
-
-        This function performs bitwise operations to compute the CRC-32
-        checksum using the polynomial 0xEDB88320.
-
-        Args:
-            crc (int): The initial CRC value.
-
-        Returns:
-            int: The updated CRC-32 checksum.
-        """
-        for _ in range(8):
-            if crc & 1:
-                crc = (crc >> 1) ^ 0xEDB88320
-            else:
-                crc >>= 1
-        return crc & 0xFFFFFFFF
-
-    @classmethod
-    def _init_crctable(cls) -> None:
-        """Initialize the CRC table if it hasn't been initialized yet."""
-        if not cls._crctable:
-            cls._crctable = [cls._gen_crc(i) for i in range(256)]
-
-    def __init__(self, password: bytes):
-        """Initialize the decrypter with the given password.
-
-        Args:
-            password (bytes): The password used to initialize decryption keys.
-        """
-        # Initialize the CRC table
-        self._init_crctable()
-
-        # Initialize keys
-        self.key0: int = 0x12345678
-        self.key1: int = 0x23456789
-        self.key2: int = 0x34567890
-
-        # Update keys with the password
-        for char in password:
-            self._update_keys(char)
-
-    def _compute_crc32_byte(self, ch: int, crc: int) -> int:
-        """Compute the CRC32 primitive on one byte.
-
-        Args:
-            ch (int): The byte to apply CRC on.
-            crc (int): The current CRC value.
-
-        Returns:
-            int: The updated CRC value after applying the byte.
-        """
-        return (crc >> 8) ^ self._crctable[(crc ^ ch) & 0xFF]
-
-    def _update_keys(self, byte_value: int) -> None:
-        """Update internal keys based on the provided byte.
-
-        Args:
-            byte_value (int): The byte used to update the decryption keys.
-        """
-        self.key0 = self._compute_crc32_byte(byte_value, self.key0)
-        self.key1 = (self.key1 + (self.key0 & 0xFF)) & 0xFFFFFFFF
-        self.key1 = (self.key1 * 134775813 + 1) & 0xFFFFFFFF
-        self.key2 = self._compute_crc32_byte((self.key1 >> 24) & 0xFF, self.key2)
-
-    def decrypt(self, data: bytes) -> bytes:
-        """Decrypt a bytes object using the initialized keys.
-
-        Args:
-            data (bytes): The encrypted data to decrypt.
-
-        Returns:
-            bytes: The decrypted data.
-        """
-        result = bytearray()
-        for byte in data:
-            temp = self.key2 | 2
-            decrypted_byte = byte ^ ((temp * (temp ^ 1)) >> 8) & 0xFF
-            self._update_keys(decrypted_byte)
-            result.append(decrypted_byte)
-        return bytes(result)
-
-
-class LZMADecompressor:
-    """
-    A custom LZMA decompressor class for decompressing data in chunks with support
-    for raw LZMA format.
-
-    Attributes:
-        eof (bool): Indicates if the end of the file has been reached during decompression.
-    """
-
-    def __init__(self) -> None:
-        """
-        Initializes the LZMADecompressor with internal states for decompression.
-
-        Raises:
-            RuntimeError: If lzma module is not available.
-        """
-        if not LZMA_AVAILABLE:
-            raise RuntimeError("LZMA module is not available. Ensure lzma is installed.")
-
-        self._decomp: Optional[lzma.LZMADecompressor] = None
-        self._unconsumed: bytes = b""
-        self.eof: bool = False
-
-    def decompress(self, data: bytes) -> bytes:
-        """
-        Decompresses the provided LZMA compressed data.
-
-        Args:
-            data (bytes): The data to be decompressed.
-
-        Returns:
-            bytes: The decompressed data.
-
-        Notes:
-            This method initializes the decompressor on the first call by reading
-            filter properties from the initial bytes of data. It continues to
-            decompress data in chunks until the end of the file.
-
-        Raises:
-            RuntimeError: If decompression is attempted when lzma is not available.
-        """
-        if not LZMA_AVAILABLE:
-            raise RuntimeError("LZMA module is not available. Decompression cannot proceed.")
-
-        if self._decomp is None:
-            self._unconsumed += data
-            if len(self._unconsumed) <= 4:
-                return b""
-            (psize,) = struct.unpack("<H", self._unconsumed[2:4])
-            if len(self._unconsumed) <= 4 + psize:
-                return b""
-
-            filter_props = self._unconsumed[4 : 4 + psize]
-            self._decomp = lzma.LZMADecompressor(
-                lzma.FORMAT_RAW,
-                filters=[{"id": lzma.FILTER_LZMA1, "props": filter_props}],
-            )
-            data = self._unconsumed[4 + psize :]
-            del self._unconsumed
-
-        result = self._decomp.decompress(data)
-        self.eof = self._decomp.eof
-        return result
-
-
-compressor_names = {
-    0: "store",
-    1: "shrink",
-    2: "reduce",
-    3: "reduce",
-    4: "reduce",
-    5: "reduce",
-    6: "implode",
-    7: "tokenize",
-    8: "deflate",
-    9: "deflate64",
-    10: "implode",
-    12: "bzip2",
-    14: "lzma",
-    18: "terse",
-    19: "lz77",
-    97: "wavpack",
-    98: "ppmd",
-}
-
-
-def _check_compression(compression: int) -> None:
-    """
-    Checks if the required compression module is available for the given compression type.
-
-    Args:
-        compression (int): The compression type identifier.
-
-    Raises:
-        RuntimeError: If the required compression module is missing.
-        NotImplementedError: If the compression type is not supported.
-    """
-    if compression == ZIP_STORED:
-        pass
-    elif compression == ZIP_DEFLATED:
-        if not zlib:
-            raise RuntimeError("Compression requires the (missing) zlib module")
-    elif compression == ZIP_BZIP2:
-        if not bz2:
-            raise RuntimeError("Compression requires the (missing) bz2 module")
-    elif compression == ZIP_LZMA:
-        if not lzma:
-            raise RuntimeError("Compression requires the (missing) lzma module")
-    else:
-        raise NotImplementedError("That compression method is not supported")
-
-
-# Use `Any` as a fallback for the zlib decompressor type to support fallback to binascii
-DecompressorType = Optional[Union[Any, bz2.BZ2Decompressor, LZMADecompressor]]
-
-
-def _get_decompressor(compress_type: int) -> DecompressorType:
-    """
-    Returns a decompressor object based on the specified compression type.
-
-    Args:
-        compress_type (int): The compression type identifier.
-
-    Returns:
-        DecompressorType:
-            The decompressor object for the specified compression type,
-            or None if no decompression is needed.
-
-    Raises:
-        NotImplementedError: If the compression type is not supported.
-        RuntimeError: If the required compression module (e.g., zlib for ZIP_DEFLATED) is
-        not available.
-    """
-    _check_compression(compress_type)
-
-    if compress_type == ZIP_STORED:
-        return None
-    elif compress_type == ZIP_DEFLATED:
-        if not ZLIB_AVAILABLE:
-            raise RuntimeError(
-                "Compression type ZIP_DEFLATED requires the zlib module, which is not available."
-            )
-        return zlib.decompressobj(-15)
-    elif compress_type == ZIP_BZIP2:
-        return bz2.BZ2Decompressor()
-    elif compress_type == ZIP_LZMA:
-        return LZMADecompressor()
-    else:
-        descr = compressor_names.get(compress_type)
-        if descr:
-            raise NotImplementedError(f"compression type {compress_type} ({descr})")
-        else:
-            raise NotImplementedError(f"compression type {compress_type}")
-
-
 class _SharedFile:
     """
     A class for managing shared, thread-safe access to a file with controlled read and write
@@ -1145,498 +461,6 @@ class _Tellable:
     def close(self) -> None:
         """Closes the underlying file-like object."""
         self.fp.close()
-
-
-class ZipExtFile(io.BufferedIOBase):
-    """
-    File-like object for reading a file entry within a ZIP archive.
-    Is returned by ZipFile.open().
-
-    This class enables reading data from compressed ZIP archive members,
-    managing decompression, CRC checking, optional decryption, and
-    seeking within the compressed data if supported by the underlying file object.
-
-    Attributes:
-        MAX_N (int): Maximum size supported by decompressor.
-        MIN_READ_SIZE (int): Block size for reading compressed data.
-        MAX_SEEK_READ (int): Chunk size for reading during seek operations.
-        mode (str): Mode in which the file is opened.
-        name (str): The filename of the archive member.
-        newlines (Optional[str]): Line separator used if text mode is specified.
-    """
-
-    # Max size supported by decompressor.
-    MAX_N = 1 << 31 - 1
-
-    # Read from compressed files in 4k blocks.
-    MIN_READ_SIZE = 4096
-
-    # Chunk size to read during seek
-    MAX_SEEK_READ = 1 << 24
-
-    def __init__(
-        self,
-        fileobj: io.BufferedIOBase,
-        mode: str,
-        zipinfo: ZipInfo,
-        pwd: Optional[bytes] = None,
-        close_fileobj: bool = False,
-    ) -> None:
-        """
-        Initializes a ZipExtFile instance for reading a compressed archive member.
-
-        Args:
-            fileobj (io.BufferedIOBase): The file object for reading the ZIP file.
-            mode (str): The mode in which the file is opened ('r' for reading).
-            zipinfo: The ZipInfo object with details about the archive member.
-            pwd (Optional[bytes]): The password for decrypting the archive member (if needed).
-            close_fileobj (bool): If True, closes the file object when the instance is closed.
-        """
-        self._fileobj = fileobj
-        self._pwd = pwd
-        self._close_fileobj = close_fileobj
-
-        self._compress_type = zipinfo.compress_type
-        self._compress_left = zipinfo.compress_size
-        self._left = zipinfo.file_size
-
-        # Initialize decompressor based on compression type
-        self._decompressor = _get_decompressor(self._compress_type)
-
-        self._eof = False
-        self._readbuffer = b""
-        self._offset = 0
-
-        self.newlines: Optional[str] = None
-
-        self.mode: str = mode
-        self.name: str = zipinfo.filename
-
-        # Declare _expected_crc as Optional[int] to allow None
-        self._expected_crc: Optional[int] = None
-        self._running_crc: int = crc32(b"")
-
-        # Check for CRC and set initial value
-        if hasattr(zipinfo, "crc_value"):
-            self._expected_crc = zipinfo.crc_value
-        else:
-            self._expected_crc = None
-
-        # Attempt to determine if the file object supports seeking
-        self._seekable = False
-        try:
-            if fileobj.seekable():
-                self._orig_compress_start = fileobj.tell()
-                self._orig_compress_size = zipinfo.compress_size
-                self._orig_file_size = zipinfo.file_size
-                self._orig_start_crc = self._running_crc
-                self._orig_crc = self._expected_crc
-                self._seekable = True
-        except AttributeError:
-            pass
-
-        # Initialize decrypter if password-protected
-        # Define _decrypter as Optional[ZipDecrypter] to accommodate both None
-        # and ZipDecrypter instances
-        self._decrypter: Optional[ZipDecrypter] = None
-
-        if pwd:
-            if zipinfo.flag_bits & _MASK_USE_DATA_DESCRIPTOR:
-                # Use extended local headers for decryption check
-                # Ensure that _raw_time is not None before using it
-                if zipinfo._raw_time is not None:
-                    check_byte = (zipinfo._raw_time >> 8) & 0xFF
-                else:
-                    raise ValueError(
-                        "Missing _raw_time value in ZipInfo; cannot perform decryption check."
-                    )
-            else:
-                # Use CRC for decryption check otherwise
-                check_byte = (zipinfo.crc_value >> 24) & 0xFF
-            h = self._init_decrypter()
-            if h != check_byte:
-                raise RuntimeError(f"Bad password for file {zipinfo.orig_filename!r}")
-
-    def _init_decrypter(self) -> Optional[int]:
-        """
-        Initializes the decrypter for encrypted ZIP files.
-
-        This method sets up the decryption mechanism if a password (`_pwd`) is provided.
-        It reads the encryption header, which consists of the first 12 bytes in the
-        cipher stream. The first 11 bytes are random, while the 12th byte contains
-        a value used to check password correctness (either the MSB of the CRC or the
-        MSB of the file time, depending on the ZIP header type).
-
-        Returns:
-            int: The 12th byte of the decrypted header, used to check password correctness.
-
-        Raises:
-            RuntimeError: If `_pwd` or `_fileobj` is not set, and decryption is attempted.
-        """
-        if not self._pwd:
-            raise RuntimeError("Password is required for decryption.")
-        if not self._fileobj:
-            raise RuntimeError("File object is required for reading encrypted data.")
-
-        # Initialize decrypter with the provided password
-        self._decrypter = ZipDecrypter(self._pwd)
-
-        # Read the first 12 bytes from the file object as the encryption header
-        # The first 12 bytes in the cypher stream is an encryption header
-        #  used to strengthen the algorithm. The first 11 bytes are
-        #  completely random, while the 12th contains the MSB of the CRC,
-        #  or the MSB of the file time depending on the header type
-        #  and is used to check the correctness of the password.
-        header = self._fileobj.read(12)
-        self._compress_left -= 12
-
-        # Decrypt the header and return the 12th byte, used for password verification
-        decrypted_header = self._decrypter.decrypt(header)
-        return decrypted_header[11]
-
-    def __repr__(self) -> str:
-        """
-        Generates a string representation of the ZipExtFile instance.
-
-        This representation includes the class module, name, file name,
-        and compression type if the file is open. If the file is closed,
-        the representation indicates that.
-
-        Returns:
-            str: A string representation of the ZipExtFile instance.
-        """
-        result: List[str] = [f"<{self.__class__.__module__}.{self.__class__.__qualname__}"]
-        if not self.closed:
-            result.append(f" name={self.name!r}")
-            if self._compress_type != ZIP_STORED:
-                result.append(
-                    f" compress_type="
-                    f"{compressor_names.get(self._compress_type, self._compress_type)}"
-                )
-        else:
-            result.append(" [closed]")
-        result.append(">")
-        return "".join(result)
-
-    def readline(self, limit: Optional[int] = -1) -> bytes:
-        """
-        Read and return a line from the stream.
-
-        This method reads bytes from the stream until a newline byte is found
-        or until `limit` bytes have been read, whichever comes first.
-
-        Args:
-            limit (Optional[int]): The maximum number of bytes to read. If negative,
-                reads until a newline is encountered or the end of the buffer.
-
-        Returns:
-            bytes: The line read from the stream, including the newline character if found.
-        """
-
-        if limit is not None and limit < 0:
-            # Shortcut common case - newline found in buffer.
-            i = self._readbuffer.find(b"\n", self._offset) + 1
-            if i > 0:
-                line = self._readbuffer[self._offset : i]
-                self._offset = i
-                return line
-
-        return super().readline(limit)
-
-    def peek(self, n: int = 1) -> bytes:
-        """
-        Return up to `n` bytes from the buffer without advancing the position.
-
-        This method attempts to return buffered bytes up to a maximum of 512 bytes,
-        reducing allocation overhead in tight loops. If there aren’t enough bytes in
-        the buffer, it reads more data into the buffer without moving the file position.
-
-        Args:
-            n (int): The number of bytes to attempt to peek from the buffer. Defaults to 1.
-
-        Returns:
-            bytes: The bytes read from the buffer without changing the current offset.
-        """
-
-        if n > len(self._readbuffer) - self._offset:
-            chunk = self.read(n)
-            if len(chunk) > self._offset:
-                # Extend the buffer with new data and reset the offset
-                self._readbuffer = chunk + self._readbuffer[self._offset :]
-                self._offset = 0
-            else:
-                # Adjust the offset if fewer bytes were read than requested
-                self._offset -= len(chunk)
-
-        # Return up to 512 bytes to reduce allocation overhead for tight loops.
-        return self._readbuffer[self._offset : self._offset + 512]
-
-    def readable(self) -> bool:
-        """
-        Check if the file is open and readable.
-
-        Raises:
-            ValueError: If the file is closed.
-
-        Returns:
-            bool: `True` if the file is open and readable, otherwise raises an exception.
-        """
-        if self.closed:  # pylint: disable=using-constant-test
-            raise ValueError("I/O operation on closed file.")
-        return True
-
-    def read(self, n: Optional[int] = -1) -> bytes:
-        """Read and return up to `n` bytes from the stream.
-
-        If `n` is omitted, `None`, or negative, reads until EOF is reached.
-        If the file is closed, raises a `ValueError`.
-
-        Args:
-            n (Optional[int]): The number of bytes to read. If `None` or negative,
-                reads until the end of the file.
-
-        Returns:
-            bytes: The data read from the stream.
-
-        Raises:
-            ValueError: If an attempt is made to read from a closed file.
-        """
-        if self.closed:  # pylint: disable=using-constant-test
-            raise ValueError("Read from closed file.")
-
-        if n is None or n < 0:
-            # Read until EOF if n is None or negative
-            buf = self._readbuffer[self._offset :]
-            self._readbuffer = b""
-            self._offset = 0
-            while not self._eof:
-                buf += self._read1(self.MAX_N)
-            return buf
-
-        end = n + self._offset
-        if end < len(self._readbuffer):
-            # Return the data directly from the buffer if it’s already available
-            buf = self._readbuffer[self._offset : end]
-            self._offset = end
-            return buf
-
-        # Otherwise, read the remaining data needed
-        n = end - len(self._readbuffer)
-        buf = self._readbuffer[self._offset :]
-        self._readbuffer = b""
-        self._offset = 0
-        while n > 0 and not self._eof:
-            data = self._read1(n)
-            if n < len(data):
-                # If more data was read than needed, store the excess
-                self._readbuffer = data
-                self._offset = n
-                buf += data[:n]
-                break
-            buf += data
-            n -= len(data)
-
-        return buf
-
-    def _update_crc(self, newdata):
-        # Update the CRC using the given data.
-        if self._expected_crc is None:
-            # No need to compute the CRC if we don't have a reference value
-            return
-        self._running_crc = crc32(newdata, self._running_crc)
-        # Check the CRC if we're at the end of the file
-        if self._eof and self._running_crc != self._expected_crc:
-            raise BadZipFile("Bad CRC-32 for file %r" % self.name)
-
-    def read1(self, n):
-        """Read up to n bytes with at most one read() system call."""
-
-        if n is None or n < 0:
-            buf = self._readbuffer[self._offset :]
-            self._readbuffer = b""
-            self._offset = 0
-            while not self._eof:
-                data = self._read1(self.MAX_N)
-                if data:
-                    buf += data
-                    break
-            return buf
-
-        end = n + self._offset
-        if end < len(self._readbuffer):
-            buf = self._readbuffer[self._offset : end]
-            self._offset = end
-            return buf
-
-        n = end - len(self._readbuffer)
-        buf = self._readbuffer[self._offset :]
-        self._readbuffer = b""
-        self._offset = 0
-        if n > 0:
-            while not self._eof:
-                data = self._read1(n)
-                if n < len(data):
-                    self._readbuffer = data
-                    self._offset = n
-                    buf += data[:n]
-                    break
-                if data:
-                    buf += data
-                    break
-        return buf
-
-    def _read1(self, n: int) -> bytes:
-        """Read up to `n` compressed bytes with at most one `read()` system call,
-        decrypt and decompress them as needed.
-
-        This method handles unconsumed data for compressed streams and manages EOF status.
-        It also updates the CRC (Cyclic Redundancy Check) for data integrity verification.
-
-        Args:
-            n (int): The maximum number of compressed bytes to read.
-
-        Returns:
-            bytes: The decompressed and decrypted data, up to `n` bytes or until EOF.
-
-        Raises:
-            ValueError: If an unsupported compression type is encountered.
-        """
-        if self._eof or n <= 0:
-            return b""
-
-        # Read from file based on compression type.
-        if self._compress_type == ZIP_DEFLATED:
-            # Handle unconsumed data.
-            data = self._decompressor.unconsumed_tail
-            if n > len(data):
-                data += self._read2(n - len(data))
-        else:
-            data = self._read2(n)
-
-        # Process data based on compression type.
-        if self._compress_type == ZIP_STORED:
-            self._eof = self._compress_left <= 0
-        elif self._compress_type == ZIP_DEFLATED:
-            n = max(n, self.MIN_READ_SIZE)
-            data = self._decompressor.decompress(data, n)
-            self._eof = self._decompressor.eof or (
-                self._compress_left <= 0 and not self._decompressor.unconsumed_tail
-            )
-            if self._eof:
-                data += self._decompressor.flush()
-        else:
-            data = self._decompressor.decompress(data)
-            self._eof = self._decompressor.eof or self._compress_left <= 0
-
-        # Adjust data length and update internal state.
-        data = data[: self._left]
-        self._left -= len(data)
-        if self._left <= 0:
-            self._eof = True
-
-        # Update CRC for the data.
-        self._update_crc(data)
-        return data
-
-    def _read2(self, n):
-        if self._compress_left <= 0:
-            return b""
-
-        n = max(n, self.MIN_READ_SIZE)
-        n = min(n, self._compress_left)
-
-        data = self._fileobj.read(n)
-        self._compress_left -= len(data)
-        if not data:
-            raise EOFError
-
-        if self._decrypter is not None:
-            data = self._decrypter(data)
-        return data
-
-    def close(self):
-        try:
-            if self._close_fileobj:
-                self._fileobj.close()
-        finally:
-            super().close()
-
-    def seekable(self):
-        if self.closed:  # pylint: disable=using-constant-test
-            raise ValueError("I/O operation on closed file.")
-        return self._seekable
-
-    def seek(self, offset, whence=os.SEEK_SET):
-        if self.closed:  # pylint: disable=using-constant-test
-            raise ValueError("seek on closed file.")
-        if not self._seekable:
-            raise io.UnsupportedOperation("underlying stream is not seekable")
-        curr_pos = self.tell()
-        if whence == os.SEEK_SET:
-            new_pos = offset
-        elif whence == os.SEEK_CUR:
-            new_pos = curr_pos + offset
-        elif whence == os.SEEK_END:
-            new_pos = self._orig_file_size + offset
-        else:
-            raise ValueError(
-                "whence must be os.SEEK_SET (0), " "os.SEEK_CUR (1), or os.SEEK_END (2)"
-            )
-
-        if new_pos > self._orig_file_size:
-            new_pos = self._orig_file_size
-
-        if new_pos < 0:
-            new_pos = 0
-
-        read_offset = new_pos - curr_pos
-        buff_offset = read_offset + self._offset
-
-        if buff_offset >= 0 and buff_offset < len(self._readbuffer):
-            # Just move the _offset index if the new position is in the _readbuffer
-            self._offset = buff_offset
-            read_offset = 0
-        # Fast seek uncompressed unencrypted file
-        elif self._compress_type == ZIP_STORED and self._decrypter is None and read_offset > 0:
-            # disable CRC checking after first seeking - it would be invalid
-            self._expected_crc = None
-            # seek actual file taking already buffered data into account
-            read_offset -= len(self._readbuffer) - self._offset
-            self._fileobj.seek(read_offset, os.SEEK_CUR)
-            self._left -= read_offset
-            read_offset = 0
-            # flush read buffer
-            self._readbuffer = b""
-            self._offset = 0
-        elif read_offset < 0:
-            # Position is before the current position. Reset the ZipExtFile
-            self._fileobj.seek(self._orig_compress_start)
-            self._running_crc = self._orig_start_crc
-            self._expected_crc = self._orig_crc
-            self._compress_left = self._orig_compress_size
-            self._left = self._orig_file_size
-            self._readbuffer = b""
-            self._offset = 0
-            self._decompressor = _get_decompressor(self._compress_type)
-            self._eof = False
-            read_offset = new_pos
-            if self._decrypter is not None:
-                self._init_decrypter()
-
-        while read_offset > 0:
-            read_len = min(self.MAX_SEEK_READ, read_offset)
-            self.read(read_len)
-            read_offset -= read_len
-
-        return self.tell()
-
-    def tell(self):
-        if self.closed:  # pylint: disable=using-constant-test
-            raise ValueError("tell on closed file.")
-        if not self._seekable:
-            raise io.UnsupportedOperation("underlying stream is not seekable")
-        filepos = self._orig_file_size - self._left - len(self._readbuffer) + self._offset
-        return filepos
 
 
 class ZipFile:
